pip_requirements.txt
Django==1.5.1
MySQL-python==1.2.4
Pillow==1.7.8
boto==2.9.0
bpssl==1.0.2
distribute==0.6.35
django-bootstrap-pagination==0.1.10
django-icybackup==0.1.1
flake8==2.0
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.17.2
mccabe==0.2.1
mecab-python==0.996
pdfminer==20110515
pep8==1.4.5
pyflakes==0.7.2
python-dateutil==2.1
python-memcached==1.48
pytz==2013b
see==1.0.1
setproctitle==1.1.7
six==1.3.0
sorl-thumbnail==11.12
wsgiref==0.1.2
